pyproject.toml
[tool.poetry]
name = "noti.py"
version = "0.3.1"
description = "Python-based notification script to post a message to a materix server via the Matrix client/server API"
authors = ["Alex Kelly <kellya@arachnitech.com>"]
readme = "README.md"
homepage = "https://github.com/kellya/notipy"
license = "MIT"
#packages = [
#    { include = "noti_py/*.py" }
#]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"
click = "^8.0.1"
requests = "^2.26.0"
coverage = "^5.5"

[tool.poetry.dev-dependencies]
python-semantic-release = "^7.19.1"
pytest = "^6.2.4"
Sphinx = "^4.3.0"
sphinx-rtd-theme = "^1.0.0"
pylint = "^2.10.2"
coverage = "^5.5"
pyflakes = "^2.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
notipy = 'noti_py.noti:main'

[tool.semantic_release]
version_variable = [
    "noti_py/noti.py:__version__",
    "pyproject.toml:version"
]
